Cell Types: Position of J is Independent of E
In the Yee scheme, the position of current density `J`
is the same as the E-field position.
In directional splitting, the current density is defined
at a different position (and is furthermore shifted in time
by half a step).

The `AllCenteredCell` numerical cell type is renamed to
`EMFCenteredCell` to reflect that.