fix out of host memory during HDF5 dumping
For HDF5 checkpoints and normal dumps we use CUDA mapped memory to transfere the
particles to the host memory. Due to the usage of cupla and the reason that
mapped memory is not supported by cupla and alpaka we fall back to native CUDA
functions. To free mapped memory the function `cudaFreeHost`. This function is also used to free normal host memory allocated with `cudaMallocHost` which is supported by cupla.
The cupla internal macros rename `cudaFreeHost` to `cuplaFreeHost` and fall back to alpaka
functionallity to free the memory. In the case where we allocated the memory with without the knowlage of cupla cupla is trowing an error and `cudaFreeHost` will never be called.
The result in PIConGPU is that with each particle dump the host memory footprint grows until we run out of memory.

- fix broken memory freeing
- `MappedBufferIntern`
  - fix that wrong function is used to allocate mapped memory
  - add workaround to free mapped memory
  - add fallpack if an CPU accelerator is used