auto scale species communication buffer
Scale species communication buffer depending on a `123^3` cell volume
per device.

HIP-clang support

- cmake workarounds
  - PIC_EMU_ATOMICADD - emulate atomicAdd(float)7 fatser than original
  - ALPAKA_EMU_MEMCPY3D - emulate hipMemcpy3Dasync via kernel
- port mallocMC to HIP

wave atomics and mallocMC opti

- optimize mallocMC by using the `waveId`
   - add function `wave_id()`
- implement PIConGPU warp atomics

compute current workaround

When using 512 threads in the kernel compute current the kernel is
crashing or the results are wrong. It is with a possibility of 95% a
compiler bug.
Add flag to disable the workaround.

fix HIP-nvcc support

- fix compile issues with HIP-nvcc

fix feature selection

- enable hardeware rsqrt for HIP on AMD
- use memory pinning when alpaka HIP accelerator is used

fix math include for HIP